Refactor test_unit.py into smaller tests
