+++ securedrop/tests/test_unit_store.py
@@ -1,0 +1,46 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+import os
+import unittest
+import zipfile
+
+
+# Set environment variable so config.py uses a test environment
+os.environ['SECUREDROP_ENV'] = 'test'
+import config
+import store
+import common
+
+
+class TestStore(unittest.TestCase):
+
+    '''The set of tests for store.py.'''
+    def setUp(self):
+        common.shared_setup()
+
+    def tearDown(self):
+        common.shared_teardown()
+
+    def test_verify(self):
+        with self.assertRaises(store.PathException):
+            store.verify(os.path.join(config.STORE_DIR, '..', 'etc', 'passwd'))
+        with self.assertRaises(store.PathException):
+            store.verify(config.STORE_DIR + "_backup")
+
+    def test_get_zip(self):
+        sid = 'EQZGCJBRGISGOTC2NZVWG6LILJBHEV3CINNEWSCLLFTUWZJPKJFECLS2NZ4G4U3QOZCFKTTPNZMVIWDCJBBHMUDBGFHXCQ3R'
+        files = ['1-abc1-msg.gpg', '2-abc2-msg.gpg']
+        filenames = common.setup_test_docs(sid, files)
+
+        archive = zipfile.ZipFile(store.get_bulk_archive(filenames))
+
+        archivefile_contents = archive.namelist()
+
+        for archived_file, actual_file in zip(archivefile_contents, filenames):
+            actual_file_content = open(actual_file).read()
+            zipped_file_content = archive.read(archived_file)
+            self.assertEquals(zipped_file_content, actual_file_content)
+
+
+if __name__ == "__main__":
+    unittest.main(verbosity=2)
